Split aws module
